refactor: organize server functions into rpc and hooks directories
Move createServerFn exports to per-module rpc/ folders and client hooks
to user/hooks, separating callable RPC boundaries from internal server helpers.
