+++ apps/auth/rpc/logout.ts
@@ -1,6 +1,6 @@
 import { redirect } from "@tanstack/react-router"
 import { createServerFn } from "@tanstack/react-start"
-import { clearAppSession } from "@/auth/session.server"
+import { clearAppSession } from "../session.server"
 
 export const logout = createServerFn({ method: "POST" }).handler(async () => {
   await clearAppSession()
