+++ apps/permissions/index.ts
@@ -1,2 +1,2 @@
-export * from "./guards"
 export * from "./permissions.server"
+export * from "./rpc"
