+++ apps/user/index.ts
@@ -1,1 +1,3 @@
+export * from "./hooks"
+export * from "./rpc"
 export * from "./user.server"
